refactor(server): extract analytics domain services
