+++ apps/server/api/package.json
@@ -15,6 +15,7 @@
     "@genfeedai/prisma": "workspace:*",
     "@genfeedai/queue-contracts": "workspace:*",
     "@genfeedai/serializers": "workspace:*",
+    "@genfeedai/server-domain": "workspace:*",
     "@genfeedai/tools": "workspace:*",
     "@genfeedai/types": "workspace:*",
     "@genfeedai/utils": "workspace:*",
