+++ apps/server/api/src/collections/content-performance/content-performance.module.ts
@@ -2,19 +2,25 @@ import { BrandMemoryModule } from '@api/collections/brand-memory/brand-memory.mo
 import { AnalyticsSyncController } from '@api/collections/content-performance/controllers/analytics-sync.controller';
 import { ContentPerformanceController } from '@api/collections/content-performance/controllers/content-performance.controller';
 import { PerformanceSummaryController } from '@api/collections/content-performance/controllers/performance-summary.controller';
-import { AnalyticsSyncService } from '@api/collections/content-performance/services/analytics-sync.service';
 import { AttributionService } from '@api/collections/content-performance/services/attribution.service';
 import { ContentPerformanceService } from '@api/collections/content-performance/services/content-performance.service';
-import { EmailDigestService } from '@api/collections/content-performance/services/email-digest.service';
 import { OptimizationCycleService } from '@api/collections/content-performance/services/optimization-cycle.service';
-import { PerformanceSummaryService } from '@api/collections/content-performance/services/performance-summary.service';
 import { OrganizationsModule } from '@api/collections/organizations/organizations.module';
 import { PostsModule } from '@api/collections/posts/posts.module';
 import { UsersModule } from '@api/collections/users/users.module';
 import { QueuesModule } from '@api/queues/core/queues.module';
 import { BrandMemorySyncService } from '@api/services/brand-memory/brand-memory-sync.service';
 import { NotificationsModule } from '@api/services/notifications/notifications.module';
+import { NotificationsService } from '@api/services/notifications/notifications.service';
+import { PrismaModule } from '@api/shared/modules/prisma/prisma.module';
+import { PrismaService } from '@api/shared/modules/prisma/prisma.service';
+import { LoggerModule } from '@libs/logger/logger.module';
+import { LoggerService } from '@libs/logger/logger.service';
 import { forwardRef, Module } from '@nestjs/common';
+import { AnalyticsSyncService } from '@server-domain/collections/content-performance/services/analytics-sync.service';
+import { EmailDigestService } from '@server-domain/collections/content-performance/services/email-digest.service';
+import { PerformanceSummaryService } from '@server-domain/collections/content-performance/services/performance-summary.service';
+import { SERVER_DOMAIN_TOKENS } from '@server-domain/server-domain.dependencies';
 
 @Module({
   controllers: [
@@ -35,6 +41,8 @@ import { forwardRef, Module } from '@nestjs/common';
     forwardRef(() => BrandMemoryModule),
     forwardRef(() => PostsModule),
     forwardRef(() => NotificationsModule),
+    forwardRef(() => PrismaModule),
+    LoggerModule,
     forwardRef(() => QueuesModule),
     forwardRef(() => OrganizationsModule),
     forwardRef(() => UsersModule),
@@ -47,6 +55,22 @@ import { forwardRef, Module } from '@nestjs/common';
     AnalyticsSyncService,
     EmailDigestService,
     BrandMemorySyncService,
+    {
+      provide: SERVER_DOMAIN_TOKENS.brandMemorySync,
+      useExisting: BrandMemorySyncService,
+    },
+    {
+      provide: SERVER_DOMAIN_TOKENS.logger,
+      useExisting: LoggerService,
+    },
+    {
+      provide: SERVER_DOMAIN_TOKENS.notifications,
+      useExisting: NotificationsService,
+    },
+    {
+      provide: SERVER_DOMAIN_TOKENS.prisma,
+      useExisting: PrismaService,
+    },
   ],
 })
 export class ContentPerformanceModule {}
