+++ apps/server/domain/src/analytics/analytics-job.types.ts
@@ -1,0 +1,4 @@
+export type AnalyticsQueueJob<TData> = {
+  data: TData;
+  updateProgress(progress: number): Promise<void>;
+};
