+++ apps/server/domain/src/server-domain.dependencies.ts
@@ -1,0 +1,166 @@
+import type { Prisma } from '@genfeedai/prisma';
+
+export const SERVER_DOMAIN_TOKENS = {
+  credentials: 'SERVER_DOMAIN_CREDENTIALS',
+  instagram: 'SERVER_DOMAIN_INSTAGRAM',
+  linkedIn: 'SERVER_DOMAIN_LINKEDIN',
+  logger: 'SERVER_DOMAIN_LOGGER',
+  mastodon: 'SERVER_DOMAIN_MASTODON',
+  notifications: 'SERVER_DOMAIN_NOTIFICATIONS',
+  pinterest: 'SERVER_DOMAIN_PINTEREST',
+  postAnalytics: 'SERVER_DOMAIN_POST_ANALYTICS',
+  posts: 'SERVER_DOMAIN_POSTS',
+  prisma: 'SERVER_DOMAIN_PRISMA',
+  tiktok: 'SERVER_DOMAIN_TIKTOK',
+  twitter: 'SERVER_DOMAIN_TWITTER',
+  youtube: 'SERVER_DOMAIN_YOUTUBE',
+  brandMemorySync: 'SERVER_DOMAIN_BRAND_MEMORY_SYNC',
+} as const;
+
+export interface ServerDomainLogger {
+  error(message: string, trace?: unknown, context?: unknown): void;
+  log(message: string, context?: unknown): void;
+  warn(message: string, context?: unknown): void;
+}
+
+export interface ServerDomainCredentialStore {
+  findOne(query: unknown): Promise<unknown>;
+}
+
+export interface ServerDomainTwitterAnalytics {
+  getMediaAnalyticsBatch(
+    tweetIds: string[],
+    accessToken: string,
+    accessTokenSecret?: string,
+  ): Promise<Map<string, unknown>>;
+}
+
+export interface ServerDomainYouTubeAnalytics {
+  getMediaAnalyticsBatch(
+    organizationId: string,
+    brandId: string,
+    videoIds: string[],
+  ): Promise<Map<string, unknown>>;
+}
+
+export interface ServerDomainSocialAnalytics {
+  getMediaAnalytics(
+    organizationId: string,
+    brandId: string,
+    externalId: string,
+  ): Promise<{
+    clicks?: number;
+    comments?: number;
+    engagementRate?: number;
+    impressions?: number;
+    likes?: number;
+    mediaType?: string;
+    reach?: number;
+    saves?: number;
+    shares?: number;
+    views?: number;
+    [key: string]: unknown;
+  }>;
+}
+
+export interface ServerDomainPostAnalytics {
+  processInstagramAnalytics(postId: string, analytics: unknown): Promise<void>;
+  processLinkedInAnalytics(postId: string, analytics: unknown): Promise<void>;
+  processMastodonAnalytics(postId: string, analytics: unknown): Promise<void>;
+  processPinterestAnalytics(postId: string, analytics: unknown): Promise<void>;
+  processTikTokAnalytics(postId: string, analytics: unknown): Promise<void>;
+  processTwitterAnalytics(postId: string, analytics: unknown): Promise<void>;
+  processYouTubeAnalytics(postId: string, analytics: unknown): Promise<void>;
+}
+
+export interface ServerDomainPosts {
+  patch(
+    postId: string,
+    data: { isAnalyticsEnabled: boolean },
+  ): Promise<unknown>;
+}
+
+export interface ServerDomainNotifications {
+  sendEmail(email: string, subject: string, html: string): Promise<void>;
+}
+
+export interface ServerDomainBrandMemorySync {
+  detectThresholdAlerts(
+    organizationId: string,
+    brandId: string,
+  ): Promise<
+    Array<{
+      type: 'spike' | 'drop';
+      metric: 'engagementRate';
+      recentAverage: number;
+      baselineAverage: number;
+      ratio: number;
+    }>
+  >;
+  syncPostPerformance(
+    organizationId: string,
+    brandId: string,
+    postId: string,
+  ): Promise<void>;
+}
+
+export interface ServerDomainPostAnalyticsRecord {
+  brandId: string | null;
+  date: Date | string | number;
+  engagementRate: number | null;
+  platform: string | null;
+  postId: string | null;
+  totalComments: number | null;
+  totalLikes: number | null;
+  totalSaves: number | null;
+  totalShares: number | null;
+  totalViews: number | null;
+  userId: string | null;
+}
+
+export interface ServerDomainPostRecord {
+  category: string | null;
+  contentRunId: string | null;
+  creativeVersion: string | null;
+  description: string | null;
+  externalId: string | null;
+  generationId: string | null;
+  hookVersion: string | null;
+  id: string;
+  label: string | null;
+  personaId: string | null;
+  publicationDate: Date | null;
+  publishIntent: string | null;
+  scheduleSlot: string | null;
+  variantId: string | null;
+}
+
+export interface ServerDomainPrisma {
+  $queryRaw<T = unknown>(query: Prisma.Sql): Promise<T>;
+  contentPerformance: {
+    create(args: unknown): Promise<unknown>;
+    findFirst(
+      args: unknown,
+    ): Promise<{ createdAt: Date; data: unknown } | null>;
+  };
+  organization: {
+    findUnique(args: unknown): Promise<{
+      label: string | null;
+      userId: string | null;
+    } | null>;
+  };
+  post: {
+    findMany(args: unknown): Promise<ServerDomainPostRecord[]>;
+  };
+  postAnalytics: {
+    aggregate(args: unknown): Promise<{
+      _avg: {
+        engagementRate: number | null;
+      };
+    }>;
+    findMany(args: unknown): Promise<ServerDomainPostAnalyticsRecord[]>;
+  };
+  user: {
+    findUnique(args: unknown): Promise<{ email: string | null } | null>;
+  };
+}
