+++ apps/server/domain/tsconfig.json
@@ -1,84 +1,92 @@
 {
   "compilerOptions": {
-    "baseUrl": ".",
-    "declaration": false,
-    "declarationMap": false,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "incremental": true,
-    "lib": ["ES2020"],
-    "module": "commonjs",
-    "moduleResolution": "bundler",
-    "types": [],
-    "outDir": "./dist",
+    "baseUrl": "..",
+    "ignoreDeprecations": "6.0",
     "paths": {
-      "@api-types/*": ["../../../packages/api-types/src/*"],
+      "@api-root/*": ["./api/*"],
+      "@api-scripts/*": ["./api/scripts/*"],
+      "@api-test/*": ["./api/test/*"],
+      "@api-types/*": ["../../packages/api-types/src/*"],
       "@api/*": ["./api/src/*"],
+      "@billing-providers": [
+        "./api/src/common/subscriptions/billing.providers.oss.ts"
+      ],
       "@clips/*": ["./clips/src/*"],
       "@cloud-types/*": ["../../packages/types/src/*"],
+      "@config/*": ["../../packages/config/src/*"],
+      "@discord/*": ["./discord/src/*"],
+      "@files/*": ["./files/src/*"],
+      "@genfeedai-types/*": ["../../packages/types/src/*"],
       "@genfeedai/agent": ["../../packages/agent/src/index.ts"],
       "@genfeedai/agent/*": ["../../packages/agent/src/*"],
-      "@genfeedai/config": ["../../packages/config/src"],
+      "@genfeedai/config": ["../../packages/config/src/index.ts"],
       "@genfeedai/config/*": ["../../packages/config/src/*"],
-      "@config/*": ["../../packages/config/src/*"],
-      "@genfeedai/constants": ["../../packages/constants/dist/index"],
-      "@genfeedai/constants/*": ["../../packages/constants/dist/*"],
-      "@genfeedai/enums": ["../../packages/enums/dist/index"],
-      "@genfeedai/enums/*": ["../../packages/enums/dist/*"],
-      "@genfeedai/integrations": ["../../packages/integrations/src"],
+      "@genfeedai/constants": ["../../packages/constants/src/index.ts"],
+      "@genfeedai/constants/*": ["../../packages/constants/src/*"],
+      "@genfeedai/ee-harness": ["../../ee/packages/harness/src/index.ts"],
+      "@genfeedai/enums": ["../../packages/enums/src/index.ts"],
+      "@genfeedai/enums/*": ["../../packages/enums/src/*"],
+      "@genfeedai/harness": ["../../packages/harness/src/index.ts"],
+      "@genfeedai/harness/*": ["../../packages/harness/src/*"],
+      "@genfeedai/helpers": ["../../packages/helpers/src/index.ts"],
+      "@genfeedai/helpers/*": ["../../packages/helpers/src/*"],
+      "@genfeedai/integrations": ["../../packages/integrations/src/index.ts"],
       "@genfeedai/integrations/*": ["../../packages/integrations/src/*"],
       "@genfeedai/interfaces": ["../../packages/interfaces/src/index.ts"],
       "@genfeedai/interfaces/*": ["../../packages/interfaces/src/*"],
       "@genfeedai/pricing": ["../../packages/pricing/src/index.ts"],
       "@genfeedai/pricing/*": ["../../packages/pricing/src/*"],
+      "@genfeedai/prisma": ["../../packages/prisma/src/index.ts"],
+      "@genfeedai/prisma/*": ["../../packages/prisma/src/*"],
       "@genfeedai/queue-contracts": [
         "../../packages/queue-contracts/src/index.ts"
       ],
       "@genfeedai/queue-contracts/*": ["../../packages/queue-contracts/src/*"],
-      "@genfeedai/serializers": ["../../packages/serializers/src"],
+      "@genfeedai/serializers": ["../../packages/serializers/src/index.ts"],
       "@genfeedai/serializers/*": ["../../packages/serializers/src/*"],
       "@genfeedai/server-domain": ["./domain/src/index.ts"],
       "@genfeedai/server-domain/*": ["./domain/src/*"],
-      "@genfeedai/storage": ["../../packages/storage/src"],
+      "@genfeedai/storage": ["../../packages/storage/src/index.ts"],
       "@genfeedai/storage/*": ["../../packages/storage/src/*"],
-      "@genfeedai/utils": ["../../packages/utils/index.ts"],
-      "@genfeedai/utils/*": ["../../packages/utils/*"],
       "@genfeedai/types": ["../../packages/types/src/index.ts"],
       "@genfeedai/types/*": ["../../packages/types/src/*"],
-      "@genfeedai/workflow-engine": ["../../packages/workflow-engine/src"],
+      "@genfeedai/ui": ["../../packages/ui/src/index.ts"],
+      "@genfeedai/ui/*": ["../../packages/ui/src/*"],
+      "@genfeedai/utils": ["../../packages/utils/index.ts"],
+      "@genfeedai/utils/*": ["../../packages/utils/*"],
+      "@genfeedai/workflow-engine": [
+        "../../packages/workflow-engine/src/index.ts"
+      ],
       "@genfeedai/workflow-engine/*": ["../../packages/workflow-engine/src/*"],
-      "@genfeedai/workflow-saas": ["../../packages/workflow-saas/src"],
+      "@genfeedai/workflow-saas": ["../../packages/workflow-saas/src/index.ts"],
       "@genfeedai/workflow-saas/*": ["../../packages/workflow-saas/src/*"],
       "@genfeedai/workflows": ["../../packages/workflows/src/index.ts"],
       "@genfeedai/workflows/*": ["../../packages/workflows/src/*"],
-      "@discord/*": ["./discord/src/*"],
-      "@fanvue/*": ["./fanvue/src/*"],
-      "@files/*": ["./files/src/*"],
       "@helpers/*": ["../../packages/helpers/src/*"],
       "@images/*": ["./images/src/*"],
       "@libs/*": ["../../packages/libs/*"],
-      "@llm/*": ["./llm/src/*"],
       "@mcp/*": ["./mcp/src/*"],
       "@notifications/*": ["./notifications/src/*"],
       "@serializers/*": ["../../packages/serializers/src/*"],
       "@server-domain/*": ["./domain/src/*"],
       "@slack/*": ["./slack/src/*"],
       "@telegram/*": ["./telegram/src/*"],
-      "@twitch/*": ["./twitch/src/*"],
+      "@ui/*": ["../../packages/ui/src/*"],
       "@videos/*": ["./videos/src/*"],
       "@voices/*": ["./voices/src/*"],
       "@workers/*": ["./workers/src/*"],
       "@workflow-engine/*": ["../../packages/workflow-engine/src/*"],
       "@workflow-saas/*": ["../../packages/workflow-saas/src/*"]
     },
-    "removeComments": true,
-    "skipLibCheck": true,
-    "sourceMap": true,
-    "strict": true,
-    "strictBindCallApply": true,
-    "strictNullChecks": true,
-    "target": "es2020"
+    "types": ["node", "vitest/globals"]
   },
-  "exclude": ["node_modules", "dist"],
-  "extends": "../../tsconfig.server.decorators.json"
+  "exclude": [
+    "node_modules",
+    "dist",
+    "**/*.spec.ts",
+    "**/*.test.ts",
+    "**/*.e2e-spec.ts"
+  ],
+  "extends": "../tsconfig.typecheck.base.json",
+  "include": ["src/**/*"]
 }
