+++ apps/server/domain/tsconfig.typecheck.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./tsconfig.json"
+}
