+++ apps/server/tsconfig.typecheck.base.json
@@ -44,6 +44,8 @@
       "@genfeedai/prisma/*": ["../../packages/prisma/src/*"],
       "@genfeedai/serializers": ["../../packages/serializers/src/index.ts"],
       "@genfeedai/serializers/*": ["../../packages/serializers/src/*"],
+      "@genfeedai/server-domain": ["./domain/src/index.ts"],
+      "@genfeedai/server-domain/*": ["./domain/src/*"],
       "@genfeedai/storage": ["../../packages/storage/src/index.ts"],
       "@genfeedai/storage/*": ["../../packages/storage/src/*"],
       "@genfeedai/types": ["../../packages/types/src/index.ts"],
@@ -66,6 +68,7 @@
       "@mcp/*": ["./mcp/src/*"],
       "@notifications/*": ["./notifications/src/*"],
       "@serializers/*": ["../../packages/serializers/src/*"],
+      "@server-domain/*": ["./domain/src/*"],
       "@slack/*": ["./slack/src/*"],
       "@telegram/*": ["./telegram/src/*"],
       "@ui/*": ["../../packages/ui/src/*"],
