+++ apps/server/workers/package.json
@@ -8,6 +8,7 @@
     "@genfeedai/interfaces": "workspace:*",
     "@genfeedai/prisma": "workspace:*",
     "@genfeedai/queue-contracts": "workspace:*",
+    "@genfeedai/server-domain": "workspace:*",
     "@genfeedai/workflows": "workspace:*",
     "@nestjs/event-emitter": "3.1.0"
   },
