+++ apps/server/workers/src/processors/api/queues/email-digest/email-digest.processor.spec.ts
@@ -1,6 +1,6 @@
-import { EmailDigestService } from '@api/collections/content-performance/services/email-digest.service';
 import { LoggerService } from '@libs/logger/logger.service';
 import { Test } from '@nestjs/testing';
+import { EmailDigestService } from '@server-domain/collections/content-performance/services/email-digest.service';
 import { EmailDigestProcessor } from '@workers/processors/api/queues/email-digest/email-digest.processor';
 import { vi } from 'vitest';
