+++ apps/server/workers/tsconfig.app.json
@@ -44,6 +44,8 @@
       ],
       "@genfeedai/serializers": ["../../../packages/serializers/src/index.ts"],
       "@genfeedai/serializers/*": ["../../../packages/serializers/src/*"],
+      "@genfeedai/server-domain": ["../domain/src/index.ts"],
+      "@genfeedai/server-domain/*": ["../domain/src/*"],
       "@genfeedai/storage": ["../../../packages/storage/src"],
       "@genfeedai/storage/*": ["../../../packages/storage/src/*"],
       "@genfeedai/types": ["../../../packages/types/src/index.ts"],
@@ -69,6 +71,7 @@
       "@mcp/*": ["../mcp/src/*"],
       "@notifications/*": ["../notifications/src/*"],
       "@serializers/*": ["../../../packages/serializers/src/*"],
+      "@server-domain/*": ["../domain/src/*"],
       "@slack/*": ["../slack/src/*"],
       "@telegram/*": ["../telegram/src/*"],
       "@twitch/*": ["../twitch/src/*"],
